Struct string_interner::backend::BucketBackend

pub struct BucketBackend<S = DefaultSymbol> { /* fields omitted */ }

An interner backend that reduces memory allocations by using string buckets.

Note

Implementation inspired by matklad’s blog post that can be found here: https://matklad.github.io/2020/03/22/fast-simple-rust-interner.html

Usage Hint

Use when deallocations or copy overhead is costly or when interning of static strings is especially common.

Usage

Fill: Efficiency of filling an empty string interner.

Resolve: Efficiency of interned string look-up given a symbol.

Allocations: The number of allocations performed by the backend.

Footprint: The total heap memory consumed by the backend.

Contiguous: True if the returned symbols have contiguous values.

Rating varies between bad, ok, good and best.

Scenario	Rating

Fill	good

Resolve	ok

Allocations	good

Footprint	ok

Supports get_or_intern_static	yes

Send + Sync	yes

Contiguous	yes