Struct string_interner::backend::BufferBackend

pub struct BufferBackend<S = DefaultSymbol> { /* private fields */ }

An interner backend that appends all interned string information in a single buffer.

Usage Hint

Use this backend if memory consumption is what matters most to you. Note though that unlike all other backends symbol values are not contigous!

Usage

Fill: Efficiency of filling an empty string interner.

Resolve: Efficiency of interned string look-up given a symbol.

Allocations: The number of allocations performed by the backend.

Footprint: The total heap memory consumed by the backend.

Contiguous: True if the returned symbols have contiguous values.

Iteration: Efficiency of iterating over the interned strings.

Rating varies between bad, ok, good and best.

Scenario	Rating

Fill	best

Resolve	bad

Allocations	best

Footprint	best

Supports get_or_intern_static	no

Send + Sync	yes

Contiguous	no

Iteration	bad