Struct string_interner::StringInterner

pub struct StringInterner<S, H = RandomState> where
    S: Symbol,
    H: BuildHasher,  { /* fields omitted */ }

Caches strings efficiently, with minimal memory footprint and associates them with unique symbols. These symbols allow constant time comparisons and look-ups to the underlying interned strings.

Methods

impl<S> StringInterner<S> where
S: Symbol,

pub fn new() -> StringInterner<S, RandomState>

Creates a new empty StringInterner.

pub fn with_capacity(cap: usize) -> Self

Creates a new StringInterner with the given initial capacity.

pub fn capacity(&self) -> usize

Returns the number of elements the StringInterner can hold without reallocating.

pub fn reserve(&mut self, additional: usize)

Reserves capacity for at least additional more elements to be interned into self.

The collection may reserve more space to avoid frequent allocations. After calling reserve, capacity will be greater than or equal to self.len() + additional. Does nothing if capacity is already sufficient.

impl<S, H> StringInterner<S, H> where
S: Symbol,
H: BuildHasher,

pub fn with_hasher(hash_builder: H) -> StringInterner<S, H>

Creates a new empty StringInterner with the given hasher.

pub fn with_capacity_and_hasher(
cap: usize,
hash_builder: H
) -> StringInterner<S, H>

Creates a new empty StringInterner with the given initial capacity and the given hasher.

pub fn get_or_intern<'a, T>(&mut self, val: T) -> S where
T: Into<Cow<'a, str>>,

Interns the given value.

Returns a symbol to access it within this interner.

This either copies the contents of the string (e.g. for str) or moves them into this interner (e.g. for String).

pub fn resolve(&self, symbol: S) -> Option<&str>

Returns the string slice associated with the given symbol if available, otherwise returns None.

pub unsafe fn resolve_unchecked(&self, symbol: S) -> &str

Returns the string associated with the given symbol.

Note

This does not check whether the given symbol has an associated string for the given string interner instance.

Safety

This will result in undefined behaviour if the given symbol had no associated string for this interner instance.

pub fn get<T>(&self, val: T) -> Option<S> where
T: AsRef<str>,

Returns the symbol associated with the given string for this interner if existent, otherwise returns None.

pub fn len(&self) -> usize

Returns the number of uniquely interned strings within this interner.

pub fn is_empty(&self) -> bool

Returns true if the string interner holds no elements.

Important traits for Iter<'a, S>
impl<'a, S> Iterator for Iter<'a, S> where
S: Symbol + 'a, type Item = (S, &'a str);

pub fn iter(&self) -> Iter<S>

Returns an iterator over the interned strings.

Important traits for Values<'a, S>
impl<'a, S> Iterator for Values<'a, S> where
S: Symbol + 'a, type Item = &'a str;

pub fn iter_values(&self) -> Values<S>

Returns an iterator over all intern indices and their associated strings.

pub fn shrink_to_fit(&mut self)

Shrinks the capacity of the interner as much as possible.