Trait string_interner::symbol::Symbol

pub trait Symbol: Copy + Eq {
    fn try_from_usize(index: usize) -> Option<Self>;

    fn to_usize(self) -> usize;
}

Types implementing this trait can be used as symbols for string interners.

The StringInterner::get_or_intern method returns Symbol types that allow to look-up the original string using StringInterner::resolve.

Note

Optimal symbols allow for efficient comparisons and have a small memory footprint.

Required methods

fn try_from_usize(index: usize) -> Option<Self>

Creates a symbol from a usize.

Returns None if index is out of bounds for the symbol.

fn to_usize(self) -> usize

Returns the usize representation of self.