Module string_interner::backend

Backends for the StringInterner.

The backend is the method or strategy that handles the actual interning. There are trade-offs for the different kinds of backends. A user should find the backend that suits their use case best.

Structs

BucketBackend

An interner backend that reduces memory allocations by using string buckets.

BufferBackend

An interner backend that appends all interned string information in a single buffer.

SimpleBackend

A simple backend that stores a separate allocation for every interned string.

StringBackend

An interner backend that accumulates all interned string contents into one string.

Traits

Backend

Types implementing this trait may act as backends for the string interner.

Type Definitions

DefaultBackend

The default backend recommended for general use.