Trait string_interner::backend::Backend

pub trait Backend: Default {
    type Symbol: Symbol;
    fn with_capacity(cap: usize) -> Self;

    fn intern(&mut self, string: &str) -> Self::Symbol;

    fn shrink_to_fit(&mut self);

    fn resolve(&self, symbol: Self::Symbol) -> Option<&str>;

    unsafe fn resolve_unchecked(&self, symbol: Self::Symbol) -> &str;

    fn intern_static(&mut self, string: &'static str) -> Self::Symbol { ... }
}

Types implementing this trait may act as backends for the string interner.

The job of a backend is to actually store, manage and organize the interned strings. Different backends have different trade-offs. Users should pick their backend with hinsight of their personal use-case.

Associated Types

type Symbol: Symbol

The symbol used by the string interner backend.

Required methods

fn with_capacity(cap: usize) -> Self

Creates a new backend for the given capacity.

The capacity denotes how many strings are expected to be interned.

fn intern(&mut self, string: &str) -> Self::Symbol

Interns the given string and returns its interned ref and symbol.

Note

The backend must make sure that the returned symbol maps back to the original string in its resolve method.

fn shrink_to_fit(&mut self)

Shrink backend capacity to fit interned symbols exactly.

fn resolve(&self, symbol: Self::Symbol) -> Option<&str>

Resolves the given symbol to its original string contents.

unsafe fn resolve_unchecked(&self, symbol: Self::Symbol) -> &str

Resolves the given symbol to its original string contents.

Safety

Does not perform validity checks on the given symbol and relies on the caller to be provided with a symbol that has been generated by the intern or intern_static methods of the same interner backend.

Provided methods

fn intern_static(&mut self, string: &'static str) -> Self::Symbol

Interns the given static string and returns its interned ref and symbol.

Note

The backend must make sure that the returned symbol maps back to the original string in its resolve method.