Struct string_interner::StringInterner

pub struct StringInterner<B = DefaultBackend<DefaultSymbol>, H = DefaultHashBuilder> where
    B: Backend,
    H: BuildHasher,  { /* fields omitted */ }

Data structure to intern and resolve strings.

Caches strings efficiently, with minimal memory footprint and associates them with unique symbols. These symbols allow constant time comparisons and look-ups to the underlying interned strings.

The following API covers the main functionality:

StringInterner::get_or_intern: To intern a new string.

This maps from string type to symbol type.

StringInterner::resolve: To resolve your already interned strings.

This maps from symbol type to string type.

Implementations

impl<B, H> StringInterner<B, H> where
B: Backend,
<B as Backend>::Symbol: Symbol,
H: BuildHasher + Default,

pub fn new() -> Self

Creates a new empty StringInterner.

pub fn with_capacity(cap: usize) -> Self

Creates a new StringInterner with the given initial capacity.

impl<B, H> StringInterner<B, H> where
B: Backend,
<B as Backend>::Symbol: Symbol,
H: BuildHasher,

pub fn with_hasher(hash_builder: H) -> Self

Creates a new empty StringInterner with the given hasher.

pub fn with_capacity_and_hasher(cap: usize, hash_builder: H) -> Self

Creates a new empty StringInterner with the given initial capacity and the given hasher.

pub fn len(&self) -> usize

Returns the number of strings interned by the interner.

pub fn is_empty(&self) -> bool

Returns true if the string interner has no interned strings.

pub fn get<T>(&self, string: T) -> Option<<B as Backend>::Symbol> where
T: AsRef<str>,

Returns the symbol for the given string if any.

Can be used to query if a string has already been interned without interning.

pub fn get_or_intern<T>(&mut self, string: T) -> <B as Backend>::Symbol where
T: AsRef<str>,

Interns the given string.

Returns a symbol for resolution into the original string.

Panics

If the interner already interns the maximum number of strings possible by the chosen symbol type.

pub fn get_or_intern_static(
&mut self,
string: &'static str
) -> <B as Backend>::Symbol

Interns the given 'static string.

Returns a symbol for resolution into the original string.

Note

This is more efficient than StringInterner::get_or_intern since it might avoid some memory allocations if the backends supports this.

Panics

If the interner already interns the maximum number of strings possible by the chosen symbol type.

pub fn shrink_to_fit(&mut self)

Shrink backend capacity to fit the interned strings exactly.

pub fn resolve(&self, symbol: <B as Backend>::Symbol) -> Option<&str>

Returns the string for the given symbol if any.