Module string_interner::symbol

Interfaces and types to be used as symbols for the StringInterner.

The StringInterner::get_or_intern method returns Symbol types that allow to look-up the original string using StringInterner::resolve.

Structs

SymbolU16

Symbol that is 16-bit in size.

SymbolU32

Symbol that is 32-bit in size.

SymbolUsize

Symbol that is the same size as a pointer (usize).

Traits

Symbol

Types implementing this trait can be used as symbols for string interners.

Type Aliases

DefaultSymbol

The symbol type that is used by default.